Restructured jqm-model to allow the creation of bundles and turned the hsql adapter into one
