+++ jqm-all/jqm-model/jqm-core/src/main/java/com/enioka/jqm/jdbc/package-info.java
