+++ jqm-all/jqm-model/jqm-core/src/main/java/com/enioka/jqm/model/Message.java
