+++ jqm-all/jqm-model/jqm-core/src/main/resources/sql/mysql.sql
