+++ jqm-all/jqm-model/jqm-impl-hsql/src/main/java/com/enioka/jqm/implhsql/DbImplHsql.java
@@ -1,4 +1,4 @@
-package com.enioka.jqm.jdbc;
+package com.enioka.jqm.implhsql;
 
 import java.sql.Connection;
 import java.sql.DatabaseMetaData;
@@ -7,6 +7,8 @@ import java.sql.SQLException;
 import java.util.List;
 import java.util.Properties;
 
+import com.enioka.jqm.jdbc.DbAdapter;
+import com.enioka.jqm.jdbc.DbConn;
 import com.enioka.jqm.model.JobInstance;
 import com.enioka.jqm.model.Queue;
